+++ library/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Add `input` of schema to `getter` function of `recursive` and `recursiveAsync` schema (pull request #441)
+
 ## v0.28.1 (February 06, 2024)
 
 - Fix bug in `union` and `unionAsync` schema for transformed inputs (issue #420)